Feat: 온보딩 - 닉네임 변경 step 컴포넌트 분리
